Add files via upload
First attempt: 
Using HuggingFace API and GPT2 model. 
Plus: The program is accepting PDF upload, extracting the file, reading the file, and displaying it's contents back to the user. 
Minus: The song generated from the PDF has nothing to do with the contents of the PDF. Ex. A PDF upload on physics notes turns into a long excerpt of the history of IBM.